Updated when getting similar animes and when adding episodes
Now when you get a similar anime, now the API search a anime with one of the genres that the given anime has, instead of searching just using the first genre of the given anime, at the same time, there were added many verification. Also, when adding a new episode to an anime, there are many verification. And the README was updated with more explanations in the needed things to run the API