Changes
There has been a lot of changes. The database has been changed from sqlite to mongodb it changing the way the api manages the data, and there has been added a higher level of security using bcrypt and jsonWebTokens to encrypt the user data, not like before that the api didn't do nothing with the data, saving a plain passwords and sending directly the userId. There are still more things to change but more than the 70% of the api has been migrated